Move env name define to login module
