+++ src/env.rs
@@ -10,31 +10,6 @@ pub const ENV_CLEANUP_USERS: &str = "CLEANUP_USERS";
 
 pub const ENV_DATABASE_URL: &str = "DATABASE_URL";
 
-#[cfg(feature = "github-login")]
-pub const ENV_GITHUB_APP_CLIENT_ID: &str = "GITHUB_APP_CLIENT_ID";
-#[cfg(feature = "github-login")]
-pub const ENV_GITHUB_APP_CLIENT_SECRET: &str = "GITHUB_APP_CLIENT_SECRET";
-
-#[cfg(feature = "gitlab-login")]
-pub const ENV_GITLAB_APP_CLIENT_ID: &str = "GITLAB_APP_CLIENT_ID";
-#[cfg(feature = "gitlab-login")]
-pub const ENV_GITLAB_APP_CLIENT_SECRET: &str = "GITLAB_APP_CLIENT_SECRET";
-
-#[cfg(feature = "google-login")]
-pub const ENV_GOOGLE_APP_CLIENT_ID: &str = "GOOGLE_APP_CLIENT_ID";
-#[cfg(feature = "google-login")]
-pub const ENV_GOOGLE_APP_CLIENT_SECRET: &str = "GOOGLE_APP_CLIENT_SECRET";
-
-#[cfg(feature = "microsoft-login")]
-pub const ENV_MICROSOFT_APP_CLIENT_ID: &str = "MICROSOFT_APP_CLIENT_ID";
-#[cfg(feature = "microsoft-login")]
-pub const ENV_MICROSOFT_APP_CLIENT_SECRET: &str = "MICROSOFT_APP_CLIENT_SECRET";
-
-#[cfg(feature = "third-party-login")]
-pub const ENV_STATIC_FILES_BASE_DIR: &str = "STATIC_FILES_BASE_DIR";
-#[cfg(feature = "third-party-login")]
-pub const ENV_USE_HTTPS: &str = "USE_HTTPS";
-
 pub fn init() {
     // LOAD ENV VAR from .env if dotenv feature is enable
     #[cfg(feature = "dotenv")]
@@ -44,10 +19,3 @@ pub fn init() {
 }
 
 pub use std::env::*;
-
-
-
-#[cfg(feature = "third-party-login")]
-pub fn static_files_base_dir() -> String {
-    var(ENV_STATIC_FILES_BASE_DIR).unwrap_or("./web/".to_string())
-}
