+++ src/login/tools.rs
@@ -1,7 +1,8 @@
-use crate::env;
+use crate::login::env;
+use crate::env as app_;
 
 pub fn scheme() -> String {
-    let scheme = if env::var(env::ENV_USE_HTTPS).unwrap_or(String::from("0")) == "1" {
+    let scheme = if app_::var(env::ENV_USE_HTTPS).unwrap_or(String::from("0")) == "1" {
         "https"
     }
     else {
